Move Swift-related files to "generator" package
Moved remaining files from "...platform.swift" to
"...generator.swift" package. Before the generator refactoring
project started, generator-related
classes were scattered inconsistently between these "platform"
and "generator" packages. The refactoring consolidates
everything under "generator", making "platform" obsolete.

Also renamed "SwiftNameResolver" to "SwiftNestedNamesResolver".
This is a temporary renaming to avoid name clashes with
the new LIME-based Swift generator. After the new generator is
introduced, this class will be deleted.

See: #621
