--- gluecodium/src/main/java/com/here/gluecodium/platform/swift/package-info.java
@@ -1,25 +1,0 @@
-/*
- * Copyright (C) 2016-2019 HERE Europe B.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-/**
- * Generator suite for Swift platform
- *
- * <p>Generates Swift API and bindings to BaseAPI layer
- */
-package com.here.gluecodium.platform.swift;
