+++ gluecodium/src/test/java/com/here/gluecodium/generator/swift/SwiftModelBuilderContainerTest.kt
@@ -59,7 +59,7 @@ import org.junit.runners.JUnit4
 @RunWith(JUnit4::class)
 class SwiftModelBuilderContainerTest {
     @MockK private lateinit var signatureResolver: LimeSignatureResolver
-    @MockK private lateinit var nameResolver: SwiftNameResolver
+    @MockK private lateinit var nameResolver: SwiftNestedNamesResolver
     @MockK private lateinit var typeMapper: SwiftTypeMapper
 
     private val swiftConstant = SwiftConstant("", null, SwiftType.VOID, SwiftValue(""))
